pyproject.toml
[tool.poetry]
name = "geospatial-data-analysis"
version = "0.1.0"
description = "analysis package for geospatial data"
authors = ["John Park <johnpark@johnssonslab.info>"]
readme = "README.md"
packages = [{include = "geospatial_data_analysis"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
geopandas = "^0.13.2"
pydantic = "^2.0.3"
logging = "^0.4.9.6"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pyreadstat = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
